Scripts optimization

+++ scenarios/dmake.sh
@@ -95,5 +95,10 @@ pkg_install() {
 		"$_allinstall" \
 		"installing..." \
 		"installed"
-	cp -rf $MINGW_PERL_PREFIX/share/startup $MINGW_PERL_PREFIX/bin/
+
+	if ! [ -f $BUILD_DIR/$P_V/post-install.marker ]
+	then
+		cp -rf $MINGW_PERL_PREFIX/share/startup $MINGW_PERL_PREFIX/bin/
+		touch $BUILD_DIR/$P_V/post-install.marker
+	fi
 }
